While validating my program I noticed this dds/deal bug
I reported it to the Google code project and to Thomas Andrews.

I think the hacky fix I did should fix it, and I think the SK is the
only one with this problem (likely because it is only a problem the first
time - not sure - it usually happens the first time, but might have
happened later than that once.