Script to evaluate question 1 of Ovunc on the miniNT
** First seat and SOUTH - you are holding: 943 86 AKQ1096 73

I would like to know the answers for following questions.

(1) When our partnership holds 22 to 23 points and less than 8 cards in the majors - what is the % of times we make 3NT from South?

with the following parameters

(1) North has 2-4S, 2-5H, 1-5D, and 2-6C and 13-14HCP.

The idea is the following. If you open 1NT 10-13 with this hand:

Q1 - you reach bunch of 22-23 HCP games other people may not reach. How successful will these games be?